==> 05_Figure_1_E.Rmd <==
---
title: "A novel role for alkyladenine DNA glycosylase in regulating alkylation-induced ER stress"
subtitle: "Figure 1 E"
author: "L Milano, CF Charlier, R Andreguetti, E Healing, MP Thomé, R Elliott, JY Masson, LD Samson, G Lenz, JAP Henriques, A Nohturfft and LB Meira"
output:
  html_notebook:
    toc: TRUE
    toc_float: TRUE
    toc_depth: 3
    code_folding: "hide"
    number_sections: FALSE
    theme: "readable"
    highlight: "tango"
    fig_caption: TRUE
    css: "./source/styles.css"
---

```{js}
function myFunction(id) {
  var x = document.getElementById(id);
  if (x.style.display === 'none') {
    x.style.display = 'block';
  } else {
    x.style.display = 'none';
  }
}
```

<div class="comments">
Note: This analysis requires input data generated by an earlier script.
</div>

## Setup  

* Clear memory and console  
* Load packages  
* Get current script name  


```{r CLEAR MEMORY AND PACKAGES AND CONSOLE, results="hide"}
# Clear memory
rm(list=setdiff(ls(all=TRUE), c(".Random.seed")))
# Clear console
cat("\014")
```

```{r include=FALSE, results="hide"}
library(magrittr)
library(ggplot2)
library(scales)
library(dplyr)
library(superheat)
```


```{r class.output="txt_output"}
this.script <- rstudioapi::getActiveDocumentContext()$path %>% basename
cat("Script:", this.script)
```

## Load data  
### Load mean expression data  
Includes probe IDs, gene symbols and gene names.  
Probe IDs are not unique (due to gene mapping).  

```{r class.output="txt_output"}
mean.expr.file <- "./data/01_Raw_Data_Processing.Rmd.expr_mean.txt"
stopifnot(file.exists(mean.expr.file))
matrix.df.mean <- read.table(mean.expr.file, sep="\t", header=T,
                             stringsAsFactors = FALSE) %>% 
  dplyr::select(-GENENAME) %>% 
  dplyr::arrange(SYMBOL, PROBEID) %>% 
  magrittr::set_rownames(NULL)
stopifnot(any(duplicated(matrix.df.mean$PROBEID)))
cat("Read file:", mean.expr.file)
rm(mean.expr.file)
```


### Load Xbp1 overexpression targets  

The original input file was obtained from the [enrichr](https://maayanlab.cloud/Enrichr/) website ([Chen et al., 2013](https://pubmed.ncbi.nlm.nih.gov/23586463/); [Kuleshov et al., 2016](https://pubmed.ncbi.nlm.nih.gov/27141961/)).

enrichr library: ["Gene_Perturbations_from_GEO_up"](https://maayanlab.cloud/Enrichr/#stats)  
Genesets:  

* "Xbp1 OE mouse GSE46178 sample 300"  
* "Xbp1 OE mouse GSE46178 sample 1437"  


```{r}
enrichr.file <- "./data/Gene_Perturbations_from_GEO_up.txt.gz"
stopifnot(file.exists(enrichr.file))
enr <- readLines(con=enrichr.file)
xbp1.oe <- list()
xbp1.oe$sample300 <- grep("Xbp1 OE mouse GSE46178 sample 300", enr, value=T) %>% 
  strsplit(., split="\t") %>% .[[1]] %>% .[-(1:2)]
xbp1.oe$sample1437 <- grep("Xbp1 OE mouse GSE46178 sample 1437", enr, value=T) %>% 
  strsplit(., split="\t") %>% .[[1]] %>% .[-(1:2)]
```


```{r results="hide"}
# Number of genes:
sapply(xbp1.oe, length)
```


```{r results="hide"}
# Percent overlap:  
cat(percent(sum(duplicated(unlist(xbp1.oe))) / max(sapply(xbp1.oe, length))))
```

  
```{r results="hide"}
# Combine both lists:
xbp1.oe <- sort(unique(unlist(xbp1.oe)))
cat("Number of genes selected:", length(xbp1.oe))
```


```{r}
# Map to mouse genes
ms.hu.map.file <- "./data/mouse_human_symbols_2018_03_05b.txt.gz"
stopifnot(file.exists(ms.hu.map.file))
ms.hu.map <- read.table(ms.hu.map.file, sep="\t", header=TRUE,
                        stringsAsFactors = FALSE)
```

```{r include="FALSE"}
head(ms.hu.map)
```


```{r class.output="txt_output"}
# Now map Xbp1.OE genes to mouse:
xbp1.oe.ms <- subset(ms.hu.map, human %in% xbp1.oe) %>% 
  .$mouse %>% unique %>% sort
cat("Number of mouse Xbp1.OE genes selected:", length(xbp1.oe.ms))
rm(ms.hu.map, xbp1.oe)
```


```{r include=FALSE}
head(matrix.df.mean)
```

### Load toptable (Wild-type MMS versus Ctrl)  
```{r load toptable, class.output="txt_output"}
tt.file <- "Supplemental_Data_1.xlsx"
stopifnot(file.exists(tt.file))
tt <- readxl::read_xlsx(path = tt.file, sheet = "wt.mms.vs.wt.ctrl") %>% 
  dplyr::select(PROBEID, adj.P.Val)
cat("File read:", tt.file)
rm(tt.file)
```

```{r include=FALSE, results="hide"}
head(tt)
```

### Load genes induced by MMS in wild-type liver  
```{r class.output="txt_output"}
DEG.file <- "./data/02_Suppl_Table_01.Rmd.DEgenes.RDS"
stopifnot(file.exists(DEG.file))
test.results.genes.mms <- readRDS(DEG.file)
wt.mms.up <- readRDS(DEG.file) %>% .$wt.mms.vs.wt.ctrl %>% .$up
cat("Read file:", DEG.file)
rm(DEG.file)
```

```{r class.output="txt_output"}
cat("Number of genes:", length(wt.mms.up))
```



## Prepare data frame for plotting  
### Aggregate expression matrix  
Where genes are represented by more than one probe, choose probe with lowest p value.  

* First add adjusted p-values to expression data  
* Then aggregate symbols by lowest p value  

```{r results="hide"}
expr.df <- merge(x=tt, y=matrix.df.mean, by="PROBEID", all.x=FALSE, all.y=TRUE) %>% 
  dplyr::group_by(SYMBOL) %>% 
  dplyr::slice(which.min(adj.P.Val)) %>% 
  dplyr::ungroup(.) %>% 
  dplyr::arrange(SYMBOL) %>% 
  dplyr::select(-PROBEID, -adj.P.Val) %>% 
  tidyr::drop_na()
head(expr.df)
```

```{r class.output="txt_output"}
cat("Number of rows:", comma(nrow(expr.df)))
```


### Filter data frame to select genes upregulated in response to MMS in wild-type liver  
```{r class.output="txt_output"}
df.wt.mms.up <- subset(expr.df, SYMBOL %in% wt.mms.up) %>% 
  tidyr::drop_na()
stopifnot(!any(duplicated(df.wt.mms.up$SYMBOL)))
cat("Rows selected:", nrow(df.wt.mms.up))
```

### Further filter data frame on Xbp1 overexpression targets  
```{r class.output="txt_output"}
df.wt.mms.up.xbp1 <- subset(df.wt.mms.up, SYMBOL %in% xbp1.oe.ms) %>%
  tidyr::drop_na()
cat("Number of rows selected:", nrow(df.wt.mms.up.xbp1))
```


### Calculate log2 fold changes  
```{r results="hide"}
# Convert data frame to matrix
expr.mx <- df.wt.mms.up.xbp1 %>% 
  tibble::column_to_rownames("SYMBOL") %>% 
  dplyr::select(Wt_Ctrl,Ko_Ctrl, Wt_MMS, Ko_MMS) %>% 
  as.matrix(.)
head(expr.mx, 4)
```


```{r calculate fc}
fold.changes <- expr.mx - expr.mx[,1]
fold.changes <- fold.changes[order(fold.changes[,"Wt_MMS"], decreasing=TRUE),]
fold.changes
```


## Plot heatmap  
```{r}
# Calculate mid point for heat.pal.values argument to superheat function
zero <- (0 - min(fold.changes)) / ((0 - min(fold.changes)) + max(fold.changes))
```


```{r superheat, fig.height=7, fig.width=7}
superheat(fold.changes, heat.pal = c("navy", "white", "firebrick3"),
          heat.pal.values = c(0, zero, 1)
          )
```

```{r results="asis"}
out.file.heatmap <- "Figure_1_E.pdf"

pdf(file=out.file.heatmap)
superheat(fold.changes, heat.pal = c("navy", "white", "firebrick3"),
          heat.pal.values = c(0, zero, 1)
          )
dev.off()
```

```{r class.output="txt_output"}
cat("Saved:", out.file.heatmap)
```


<hr>

### Session info  

<button class="button" onclick="myFunction('DIV_5')">Show/hide session info</button>
<div id="DIV_5" class="div_default_hide">

```{r SESSION INFO DATE, results="asis"}
cat("Date:", format(Sys.time(), "%a %d-%b-%Y %H:%M:%S"), "<br>\n")
```

```{r print_session_info, R.options=list(width=70)}
devtools::session_info()
```
</div>
  
```{js}
var divsToHide = document.getElementsByClassName("div_default_hide");
for(var i = 0; i < divsToHide.length; i++)
{
  divsToHide[i].style.display = 'none';
}
```
