==> 06_Figure_S1.Rmd <==
---
title: "A novel role for alkyladenine DNA glycosylase in regulating alkylation-induced ER stress"
subtitle: "Figure S1"
author: "L Milano, CF Charlier, R Andreguetti, E Healing, MP Thomé, R Elliott, JY Masson, LD Samson, G Lenz, JAP Henriques, A Nohturfft and LB Meira"
output:
  html_notebook:
    toc: TRUE
    toc_float: TRUE
    toc_depth: 3
    code_folding: "hide"
    number_sections: FALSE
    theme: "readable"
    highlight: "tango"
    fig_caption: TRUE
    css: "./source/styles.css"
---

```{js}
function myFunction(id) {
  var x = document.getElementById(id);
  if (x.style.display === 'none') {
    x.style.display = 'block';
  } else {
    x.style.display = 'none';
  }
}
```

<div class='comments'>
Note: This analysis requires input data generated by an earlier script.
</div>

## Setup  

* Clear memory and console  
* Load packages  
* Get current script name  

```{r CLEAR MEMORY AND PACKAGES AND CONSOLE, results="hide"}
# Clear memory
rm(list=setdiff(ls(all=TRUE), c(".Random.seed")))
# Clear console
cat("\014")
```

## Set up  
```{r Packages, include=FALSE, results="hide"}
library(dplyr)
library(magrittr)
library(ggplot2)
library(plotly)
```


```{r Get script name,  class.output="txt_output"}
this.script <- rstudioapi::getActiveDocumentContext()$path %>% basename
cat("Script:", this.script)
```


```{r Functions_1}
gg.get.breaks_by1 <- function(limits) {
  a <- floor(limits[1])
  b <- ceiling(limits[2])
  seq(a, b, by = 1)
}

gg.get.breaks_by2 <- function(limits) {
  a <- floor(limits[1])
  if (a %% 2 == 1) a <- a - 1
  b <- ceiling(limits[2])
  if (b %% 2 == 1) b <- b + 1
  seq(a, b, by = 2)
}
```


```{r Functions_2}
get.gg <- function(mtrx, colum.names, axis.labs, titre,
                    marker.col.false, marker.col.true="#B20000", marker.alpha=1, ttable) {

  ttable2 <- dplyr::select(ttable, PROBEID, SYMBOL, adj.P.Val) %>% 
    dplyr::group_by(PROBEID) %>%
    dplyr::slice(which.min(adj.P.Val)) %>% 
    dplyr::ungroup(.)

  df <- mtrx[,colum.names] %>% 
    magrittr::set_colnames(axis.labs) %>% 
    tibble::rownames_to_column("PROBEID")
  
  df2 <- merge(x=ttable2, y=df, by="PROBEID", all.x=FALSE, all.y=TRUE) %>% 
    dplyr::mutate(p.ok = adj.P.Val <= 0.05) %>% 
    dplyr::mutate(p.ok = as.character(p.ok)) %>% 
    dplyr::mutate(p.ok = factor(p.ok, levels=c("FALSE", "TRUE"))) %>% 
    dplyr::mutate(ID = paste0(SYMBOL, " (", PROBEID, ")")) %>% 
    dplyr::select(-PROBEID, -SYMBOL, -adj.P.Val) %>% 
    dplyr::select(ID, everything())

  res <- df2 %>% 
    ggplot(aes_string(x=axis.labs[1], y=axis.labs[2])) +
    geom_point(aes(text=ID, color=p.ok), alpha=marker.alpha) + # color=marker.col.false,
    # scale_color_discrete(name="p ≤ 0.05") +
    scale_color_manual(values=c(marker.col.false, marker.col.true), name="p ≤ 0.05") +
    scale_x_continuous(limits =c(min(mtrx), max(mtrx))) +
    scale_y_continuous(limits =c(min(mtrx), max(mtrx))) +
    geom_abline(intercept = 0, slope = 1, color="grey25") +
    geom_abline(intercept = -log2(1.75), slope = 1, color="grey25", linetype="dashed") +
    geom_abline(intercept = log2(1.75), slope = 1, color="grey25", linetype="dashed") +
    labs(title = titre)
  res
}
```

## Load data
### Load mean expression data  
Table includes probe IDs, gene symbols and gene names.  
Probe IDs and symbols are not unique (due to gene symbol mapping).  

```{r Read mean expr, class.output="txt_output"}
mean.expr.file <- "./data/01_Raw_Data_Processing.Rmd.expr_mean.txt"
matrix.df.mean <- read.table(mean.expr.file, sep="\t", header=T,
                             stringsAsFactors = FALSE) %>% 
  dplyr::select(-SYMBOL, -GENENAME) %>% 
  dplyr::distinct(PROBEID, .keep_all=TRUE) %>% 
  tibble::column_to_rownames("PROBEID")
cat("File read:", mean.expr.file)
rm(mean.expr.file)
```

```{r include=FALSE}
head(matrix.df.mean)
```


### Load differentially expressed probes (effect of MMS)  
```{r Read DE probes, class.output="txt_output"}
# list.files("data")
diff.probes.file <- "./data/02_Suppl_Table_01.Rmd.DEprobes.RDS"
stopifnot(file.exists(diff.probes.file))
test.results.probes <- readRDS(diff.probes.file)
cat("File read:", diff.probes.file)
rm(diff.probes.file)
```

```{r results="asis"}
x <- paste(utils::capture.output(str(test.results.probes)), collapse="<br>\n", sep="")
details::details(x, summary="Show summary of DE probes", lang=NULL)
rm(x)
```


### Load toptables  
```{r Read toptables, class.output="txt_output"}
# list.files()
tt.file <- "Supplemental_Data_1.xlsx"
stopifnot(file.exists(tt.file))
tt.sheet.names <- openxlsx::getSheetNames(tt.file)
# tt.sheet.names
# "ko.ctrl.vs.wt.ctrl" "ko.mms.vs.wt.mms"   "wt.mms.vs.wt.ctrl"  "ko.mms.vs.ko.ctrl"
get.xlsx.data <- function(fil, sheet) {
  openxlsx::read.xlsx(xlsxFile = fil, sheet=sheet)
}
tt.list <- lapply(tt.sheet.names, function(the.sheet) {
  res <- get.xlsx.data(fil=tt.file, sheet=the.sheet) %>% 
    dplyr::select(PROBEID, SYMBOL, logFC, adj.P.Val)
  res
}) %>% 
  set_names(tt.sheet.names)
cat("File read:", tt.file)
rm(tt.file, tt.sheet.names)
```


```{r results="asis"}
y <- paste(utils::capture.output(str(tt.list)), collapse="<br>\n", sep="")
details::details(y, summary="Show summary of toptables list", lang=NULL)
rm(y)
```


## Plot probes suppressed OR induced by MMS in wild-type OR knockout    
### Select probes to plot
```{r Select genes to plot, class.output="txt_output"}
wt.ko.down.up <- sort(unique(unlist(test.results.probes)))
stopifnot(!any(is.na(wt.ko.down.up)))
cat("Probes selected:", length(wt.ko.down.up))
```



```{r results="hide"}
# matrix.df.mean[1:4,1:4]
mx.wt.ko.down.up <- matrix.df.mean[wt.ko.down.up,]
dim(mx.wt.ko.down.up)
```


```{r warning=FALSE, message=FALSE}
gg.6.4 <- get.gg(mtrx=mx.wt.ko.down.up,
       colum.names = c("Wt_Ctrl", "Ko_Ctrl"),
       axis.labs = c("WT", "KO"),
       titre = "Control: Knockout versus Wild-type",
       marker.col.false = "#005AB5", # bp[3]
       marker.col.true = "#DC3220",
       marker.alpha=0.5,
       ttable=tt.list$ko.ctrl.vs.wt.ctrl)
gg.6.4b <- gg.6.4 + 
  theme_classic() +
  coord_equal()
ggplotly(gg.6.4b)
```


```{r class.output="txt_output"}
out.file.pdf <- "Figure_S1.pdf"
ggsave(filename=out.file.pdf, plot = gg.6.4b, width = 10, height=10, units = "cm")
cat("Saved:", out.file.pdf)
```


### Session info  

<button class="button" onclick="myFunction('DIV_1')">Show/hide session info</button>
<div id="DIV_1" class="div_default_hide">

```{r SESSION INFO DATE, results="asis"}
cat("Date:", format(Sys.time(), "%a %d-%b-%Y %H:%M:%S"), "<br>\n")
```

```{r print_session_info, R.options=list(width=70)}
devtools::session_info()
```
</div>
  
```{js}
var divsToHide = document.getElementsByClassName("div_default_hide");
for(var i = 0; i < divsToHide.length; i++)
{
  divsToHide[i].style.display = 'none';
}
```
